Update: User Registration Redirect Path
This update utilizes Devise's after_update_path_for method to handle user
redirection after updating their registration.

If the user is part of the staff roster, associated with a valid organization,
and authorized to access the organization's dashboard, they will be redirected to their staff dashboard.

If the user is an adopter or fosterer and authorized to access the adopter/foster dashboard,
they will be redirected there.

If neither condition is met, the user will be redirected to the root path.

Adds: Controller tests for user registration updates:
Test case for adopter/fosterer user redirection.
Test case for staff user redirection.