Empty state for tasks
We add an empty state for tasks that primarily uses CSS to leverage the
`:only-child` psuedo selector.

Additional changes were required for the toggling of empty state when
toggling tasks, so changes were made to the toggle-completed Stimulus
controller to add a hide class you can input from HTML, as well as
checking whether you were toggling all tasks or not to decide if we
needed the empty state to show up.